pheniqs-qiime2 pipeline

+++ mmeds/tools/tool.py
@@ -507,7 +507,6 @@ class Tool(mp.Process):
 
         submitfile = self.path / 'submitfile'
         self.add_path(submitfile, '.sh', 'submitfile')
-
         # Define the job and error files
         count = 0
         jobfile = self.path / 'jobfile_{}.lsf'.format(count)
@@ -606,42 +605,45 @@ class Tool(mp.Process):
                 if 'MMEDS_STAGE_{}'.format(i) in log_text:
                     stage = i
             self.update_doc(restart_stage=stage)
-            # Files removed
-            deleted = []
-            # TODO Move this to happen when an anlysis is restarted
-            # Go through all files in the analysis
-            for stage, files in self.stage_files.items():
-                self.logger.debug('{}: Stage: {}, Files: {}'.format(self.name, stage, files))
-                # If they should be created after the last checkpoint
-                if stage >= self.doc.restart_stage:
-                    self.logger.debug('{}: Greater than restart stage'.format(self.name))
-                    for f in files:
-                        if not f == 'jobfile' and not f == 'errorlog':
-                            deleted.append(f)
-                            # Check if they exist
-                            unfinished = self.get_file(f, True)
-                            self.logger.debug('{}: checking file {}'.format(self.name, unfinished))
-                            if unfinished.exists():
-                                self.logger.debug('{}: file exists'.format(self.name))
-                                # Otherwise delete them
-                                if unfinished.is_dir():
-                                    self.logger.debug('{}: rmtree'.format(self.name))
-                                    rmtree(unfinished)
+
+            # TODO: Find a use for this, disabled because it deletes files we want
+            if False:
+                # Files removed
+                deleted = []
+                # TODO Move this to happen when an anlysis is restarted
+                # Go through all files in the analysis
+                for stage, files in self.stage_files.items():
+                    self.logger.debug('{}: Stage: {}, Files: {}'.format(self.name, stage, files))
+                    # If they should be created after the last checkpoint
+                    if stage >= self.doc.restart_stage:
+                        self.logger.debug('{}: Greater than restart stage'.format(self.name))
+                        for f in files:
+                            if not f == 'jobfile' and not f == 'errorlog':
+                                deleted.append(f)
+                                # Check if they exist
+                                unfinished = self.get_file(f, True)
+                                self.logger.debug('{}: checking file {}'.format(self.name, unfinished))
+                                if unfinished.exists():
+                                    self.logger.debug('{}: file exists'.format(self.name))
+                                    # Otherwise delete them
+                                    if unfinished.is_dir():
+                                        self.logger.debug('{}: rmtree'.format(self.name))
+                                        rmtree(unfinished)
+                                    else:
+                                        self.logger.debug('{}: unlink'.format(self.name))
+                                        unfinished.unlink()
                                 else:
-                                    self.logger.debug('{}: unlink'.format(self.name))
-                                    unfinished.unlink()
-                            else:
-                                self.logger.debug('{}: file does exist {}'.format(self.name, unfinished))
-                else:
-                    self.logger.debug('{}: stage already passed'.format(self.name))
+                                    self.logger.debug('{}: file does exist {}'.format(self.name, unfinished))
+                    else:
+                        self.logger.debug('{}: stage already passed'.format(self.name))
 
-            # Remove the deleted files from the mongodb document for the analysis
-            finished_files = self.doc.files
-            for key in deleted:
-                del finished_files[key]
-            self.update_doc(files=finished_files)
+                # Remove the deleted files from the mongodb document for the analysis
+                finished_files = self.doc.files
+                for key in deleted:
+                    del finished_files[key]
+                self.update_doc(files=finished_files)
 
-            self.logger.debug('{}: finished file cleanup'.format(self.name))
+                self.logger.debug('{}: finished file cleanup'.format(self.name))
 
             email = ('email', self.doc.email, self.doc.owner, 'error',
                      dict(code=self.doc.access_code,
@@ -649,7 +651,9 @@ class Tool(mp.Process):
                           stage=self.doc.restart_stage,
                           study=self.doc.study_name))
             self.queue.put(email)
-            raise AnalysisError('{} failed during stage {}'.format(self.name, self.doc.restart_stage))
+            # TODO: Is this relevant?
+            if False:
+                raise AnalysisError('{} failed during stage {}'.format(self.name, self.doc.restart_stage))
 
         email = ('email', self.doc.email, self.doc.owner, 'analysis_done',
                  dict(code=self.doc.access_code,
